external function for getWETH9
