+++ src/tasks/utils.ts
@@ -1,6 +1,7 @@
 import { Contract } from "ethers";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 
+import weth9Networks from "../../node_modules/canonical-weth/networks.json";
 import { TypedDataDomain } from "../ts/ethers";
 
 export function domain(
@@ -58,4 +59,20 @@ export async function getDepositAndPlaceOrderContract({
   ).connect(ethers.provider);
 
   return authenticator;
+}
+
+export async function getWETH9Address(
+  hre: HardhatRuntimeEnvironment,
+): Promise<string> {
+  // Todo: to be refactored...
+  let weth9Address = "";
+  const chainId = (await hre.ethers.provider.getNetwork()).chainId;
+  if (chainId == 4) {
+    weth9Address = weth9Networks.WETH9["4"]["address"];
+  } else if (chainId == 1) {
+    weth9Address = weth9Networks.WETH9["1"]["address"];
+  } else if (chainId == 100) {
+    weth9Address = "0xe91D153E0b41518A2Ce8Dd3D7944Fa863463a97d";
+  }
+  return weth9Address;
 }
